Section 3 Lecture 19: Create Enum from Existing Code
Created a separate enum in the Program.cs file and added it to the Entity Data Model as an Enum Type. When adding an enum this way, the name must match the enum type exactly. Creating the enum this way means that the values are not specified in the wizard. Instead, select the Reference External Type, and then specify the statically typed reference.